Update de_DE.ts
1) Corrected wrong LUKS version in the German translation.

2) Removed Space characters before and after ,,+" to unify it with the English naming 
(also complying with the German translation for ,,LUKS1+External Header", line 1214)

3) Differing German translations for ,,External Header":
-,,Externe Kopfdaten" (line 1214)
-,,Externer Header" (line 1224)

I replaced ,,Header" with ,,Kopfdaten" as in line 1214.
The German translation uses ,,Kopfdaten" also at other places.


In my opinion ,,Kopfdaten" is not common in German. Instead ,,Header" is used.
 
Maybe it is an option to provide additional explanation in brackets ,,(Header)".

If ,,(Header)" is added as explanation, than it should be also added at all other places in the translation.